resolver: merge name_server_stats module into name_server
